refactor: :truck: co-locate AppHeader in layout
